setting vpp mock mode until root permissions issues are sorted out (#21229)

+++ nic/apollo/test/tools/setup_gtests.sh
@@ -8,6 +8,7 @@ source $CUR_DIR/../../../tools/setup_env_mock.sh $PIPELINE
 export SKIP_VERIFY=1
 export GEN_TEST_RESULTS_DIR=${BUILD_DIR}/gtest_results
 export VAL_CMD=valgrind
+export VPP_IPC_MOCK_MODE=1
 
 function finish () {
     # auto invoked on any exit
@@ -18,7 +19,9 @@ function finish () {
     sudo rm -f /tmp/*.db /tmp/pen_* /dev/shm/pds_* /dev/shm/ipc_*
     sudo rm -f /dev/shm/metrics_*
     if [ $PIPELINE == 'apulu' ]; then
-        sudo pkill -9 vpp
+        if [ -z "$VPP_IPC_MOCK_MODE" ]; then
+            sudo pkill -9 vpp
+        fi
         sudo pkill -9 dhcpd
     fi
 }
@@ -31,11 +34,13 @@ function setup () {
     ln -s ${PDSPKG_TOPDIR}/conf/${PIPELINE}/pipeline.json ${PDSPKG_TOPDIR}/conf/pipeline.json
 
     if [ $PIPELINE == 'apulu' ]; then
-        echo "Starting VPP"
-        sudo ${PDSPKG_TOPDIR}/vpp/tools/start-vpp-mock.sh --pipeline apulu
-        if [[ $? != 0 ]]; then
-            echo "Failed to bring up VPP"
-            exit -1
+        if [ -z "$VPP_IPC_MOCK_MODE" ]; then
+            echo "Starting VPP"
+            sudo ${PDSPKG_TOPDIR}/vpp/tools/start-vpp-mock.sh --pipeline apulu
+            if [[ $? != 0 ]]; then
+                echo "Failed to bring up VPP"
+                exit -1
+            fi
         fi
 
         echo "Starting dhcpd"
